Use multiplicative factors in score_manager

+++ src/state/include/state/utilities.h
@@ -27,7 +27,7 @@ enum class TerrainType { LAND, WATER, TOWER, FLAG };
 /**
  * Enum for the two game players
  */
-enum class PlayerId : int {
+enum class PlayerId : size_t {
     PLAYER1 = 0,
     PLAYER2 = 1,
     PLAYER_COUNT = 2,
